requirements.txt
aniso8601==8.0.0
appdirs==1.4.4
asgiref==3.2.10
astroid==2.4.2
backoff==1.10.0
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
colorama==0.4.3
configparser==5.0.1
cryptography==3.2.1
Cython==0.29.21
distlib==0.3.1
Django==3.1.2
et-xmlfile==1.0.1
filelock==3.0.12
Flask==1.1.2
Flask-Cors==3.0.9
Flask-RESTful==0.3.8
future==0.18.2
google==3.0.0
gunicorn==20.0.4
ibm-cloud-sdk-core==1.7.3
ibm-watson==4.7.1
idna==2.10
isort==5.5.3
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.11.2
json-excel-converter==1.3.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
Naked==0.1.31
numpy==1.19.2
oauthlib==3.1.0
opencage==1.2.1
opencv-python==4.4.0.44
openpyxl==3.0.5
packaging==20.4
pandas==1.1.3
phishing-detection==0.1.2
Pillow==7.2.0
pluggy==0.13.1
py==1.9.0
pycparser==2.20
pycryptodome==3.9.8
PyJWT==1.7.1
pylint==2.6.0
pyOpenSSL==19.1.0
pyparsing==2.4.7
PyQt5==5.15.1
PyQt5-sip==12.8.1
pyqt5-tools==5.15.1.1.7.2
pysafebrowsing==0.1.1
PySocks==1.7.1
pytesseract==0.3.6
python-dateutil==2.8.1
python-dotenv==0.14.0
python-whois==0.7.3
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
shellescape==3.8.1
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.1
tesseract==0.1.3
toml==0.10.1
tox==3.20.1
tweepy==3.9.0
unittest-xml-reporting==3.0.4
urllib3==1.25.10
virtualenv==20.1.0
websocket-client==0.48.0
Werkzeug==1.0.1
whitenoise==5.2.0
wrapt==1.12.1
xlrd==1.2.0
XlsxWriter==1.3.7
